create_discord_bot/boilerplates/requirements.txt
discord.py==1.3.4
python-dotenv